pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ballsdex-crafting"
version = "0.1.0"

description = "Crafting package for BallsDex V3"
license = "MIT"
authors = [
    { name = "Haymooed" },
]
readme = "README.md"

dependencies = [
    "django-solo>=2.2.0",
]
